requirements.txt
Flask==3.1.2
PyMySQL==1.1.1
python-dotenv==1.0.1
WeasyPrint==61.2
gunicorn==21.2.0
Werkzeug>=3.1.0
